UBI: Use an atomic64_t field for ->global_sqnum
Use an atomic64_t field for ->global_sqnum, this way we don't need to take
the ltree_lock (which is not really related to the global_sqnum member)
when updating this field and we may benefit from native atomic64
operations on some platforms.

The real goal here is to move the the ltree and ltree_lock objects to the
ubi_volume struct, since this has nothing to do in the ubi_device object.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>