mtd: spi-nor: Create a ->set_4byte() method
The procedure used to enable 4 byte addressing mode depends on the NOR
device, so let's provide a hook so that manufacturer specific handling
can be implemented in a sane way.

This is also an opportunity to create a per-manufacturer post SFDP
fixups function where we'll put all the manufacturer specific tweaks.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>