mtd: rawnand: Add NAND_NO_BBM flag
Add the NAND_NO_BBM flag to avoid dummy ->bad_block implementation
always returning 0.

This flag should never be used if the controller is able to access the
BBM and BBMs are preserved by the ECC layout in use.

We also get rid of the 2 implementations returning 0 and add the
NAND_NO_BBM flag instead.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>