mtd: spi-nor: Rework the ->quad_enable() selection logic
The goal is to move the quad_enable manufacturer specific init in the
<manuf>_post_sfdp_fixups() hooks.

First thing we do is add a spi_nor_quad_enable() wrapper that only
calls nor->quad_enable() when needed (when Quad opcodes are selected).
This way we can populate nor->quad_enable() without risking a switch
to Quad mode when we don't expect it.

Second thing we do is add an explicit no_quad_enable() implementation
so that params->quad_enable() is no longer NULL when
BFPT_DWORD15_QER_NONE is detected, and we can easily override any
manufacturer specific function when SFDP parsing succeeds.

And the last thing we do is move the default nor->quad_enable init in
spi_nor_scan() and the manufacturer specific init in
<manuf>_post_sfdp_fixups().

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>