mtd: spi-nor: Add a new flag to clear SW protection bits during init
Get rid of the last piece of code doing checks on SNOR_MFR to decide
what to do. We add a new SNOR_F flag to explicitly ask for a SR clear
at init time and this flag is set in spi_nor_scan() when the NOR
supports the locking feature or directly from the SST, Intel and Atmel
post-SFDP fixups hooks since NORs from those manufacturers always
require clearing the protection bits.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>