mtd: spi-nor: Add a new hook to let part specific code tweak the config
SFDP tables are sometimes wrong and we need a way to override the
config chosen by the SFDP parsing logic without discarding all of it.
We also need a way to add chip specific tweaks when the SPI NOR does
not support RDSFDP.

Add a new hook called after the SFDP parsing has taken place to deal
with such problems.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>