mtd: spi-nor: Add a post SFDP fixup hook for gd25q256
gd25q256 needs to tweak the ->quad_enable() implementation and the
->post_sfdp() fixup hook is the perfect place to do that. This way,
if we ever need to tweak more things for this flash, we won't have to
add new fields in flash_info.

We can get rid of the flash_info->quad_enable field as gd25q256 was
the only user.

Signed-off-by: Boris Brezillon <boris.brezillon@bootlin.com>