lib/sort.c: implement sort() variant taking context argument
Our list_sort() utility has always supported a context argument that
is passed through to the comparison routine. Now there's a use case
for the similar thing for sort().

This implements sort_r by simply extending the existing sort function
in the obvious way. To avoid code duplication, we want to implement
sort() in terms of sort_r(). The naive way to do that is

static int cmp_wrapper(const void *a, const void *b, const void *ctx)
{
  int (*real_cmp)(const void*, const void*) = ctx;
  return real_cmp(a, b);
}

sort(..., cmp) { sort_r(..., cmp_wrapper, cmp) }

but this would do two indirect calls for each comparison. Instead, do
as is done for the default swap functions - that only adds a cost of a
single easily predicted branch to each comparison call.

Aside from introducing support for the context argument, this also
serves as preparation for patches that will eliminate the indirect
comparison calls in common cases.

Requested-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Cc: Andrew Morton <akpm@linux-foundation.org>
---
Hi all,

Andrew, you acked the first version of this patch, but Rasmus proposed
a better solution and posted a v2. Can you review/ack this version.

Hans, Mauro, Andrew suggested to have this patch applied along with
its first user (the H264 backend of the hantro codec), so here it is.
Note that, if possible, I'd like to have this patch queued for the next
release even if the H264 bits don't get accepted as is. The rationale
here being that Rasmus told me he was planning to further improve the
sort() logic after the next -rc1 is out, and I fear his changes will
conflict with this patch, which might involve some kind synchronisation
(a topic branch) between the media maintainers and Andrew.

Let me know how you want to proceed with that.

Regards,

Boris