spi: mxic: Add external ECC engine support to the SPI-mem mode
Some SPI-NAND chips do not have a proper on-die ECC engine providing
error correction/detection. This is particularly an issue on embedded
devices with limited resources because all the computations must
happen in software, unless an external hardware engine is provided.

These external engines are new and Macronix is providing one which we
are going to use by hacking in the controller's SPI-mem helpers.

An entire subsystem has been contributed to support these engines but
for this one two functions are specific and are not (yet) part of the
API: one to check if the ECC engine must be used and the other to
trigger transfers when needed. They are exported through a specific
header.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>