mtd: nand: mxic-ecc: Support SPI pipelined syndrome mode
Introduce the support for another possible configuration: the ECC
engine may work as DMA master (pipelined) and move itself the data
to/from the NAND chip into the buffer, applying the necessary
corrections/computations on the fly. The syndrome layout is used as to
reduce the number of DMA master/slave changes in the SPI controller
driver.

Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>